feat: implementing partial students download based on selected filters
