+++ emis/src/main/java/org/saudigitus/emis/data/local/repository/DataManagerImpl.kt
@@ -8,23 +8,15 @@ import kotlinx.coroutines.flow.conflate
 import kotlinx.coroutines.flow.flow
 import kotlinx.coroutines.flow.flowOn
 import kotlinx.coroutines.withContext
-import org.dhis2.bindings.userFriendlyValue
 import org.dhis2.commons.bindings.dataElement
 import org.dhis2.commons.bindings.enrollment
 import org.dhis2.commons.date.DateUtils
 import org.dhis2.commons.network.NetworkUtils
 import org.hisp.dhis.android.core.D2
-import org.hisp.dhis.android.core.arch.repositories.scope.RepositoryScope
 import org.hisp.dhis.android.core.dataelement.DataElement
-import org.hisp.dhis.android.core.enrollment.Enrollment
-import org.hisp.dhis.android.core.event.Event
 import org.hisp.dhis.android.core.event.EventCreateProjection
-import org.hisp.dhis.android.core.trackedentity.TrackedEntityAttribute
-import org.hisp.dhis.android.core.trackedentity.TrackedEntityAttributeValue
-import org.hisp.dhis.android.core.trackedentity.TrackedEntityInstance
 import org.saudigitus.emis.data.local.DataManager
 import org.saudigitus.emis.data.local.util.SqlRaw
-import org.saudigitus.emis.data.model.Attendance
 import org.saudigitus.emis.data.model.CalendarConfig
 import org.saudigitus.emis.data.model.EMISConfig
 import org.saudigitus.emis.data.model.EMISConfigItem
@@ -37,7 +29,7 @@ import org.saudigitus.emis.service.RuleEngineRepository
 import org.saudigitus.emis.ui.attendance.AttendanceOption
 import org.saudigitus.emis.ui.components.DropdownItem
 import org.saudigitus.emis.utils.Constants
-import org.saudigitus.emis.utils.DateHelper
+import org.saudigitus.emis.utils.Transformations
 import org.saudigitus.emis.utils.Utils
 import org.saudigitus.emis.utils.Utils.getAttendanceStatusColor
 import org.saudigitus.emis.utils.eventsWithTrackedDataValues
@@ -51,11 +43,10 @@ import javax.inject.Inject
 class DataManagerImpl
 @Inject constructor(
     val d2: D2,
+    private val transformations: Transformations,
     val networkUtils: NetworkUtils,
     val ruleEngineRepository: RuleEngineRepository,
 ) : DataManager {
-
-    private lateinit var currentProgram: String
 
     private fun getAttributeOptionCombo() =
         d2.categoryModule().categoryOptionCombos()
@@ -236,7 +227,7 @@ class DataManagerImpl
         program: String,
         stage: String,
         dataElementIds: List<String>,
-        options: List<String>,
+        dataValues: List<String>,
     ): Flow<List<SearchTeiModel>> = flow {
         emit(
             d2.eventsWithTrackedDataValues(
@@ -248,7 +239,7 @@ class DataManagerImpl
                     Pair(trackedEntityDataValue.dataElement(), trackedEntityDataValue.value())
                 }
                 dataElements?.keys?.containsAll(dataElementIds) == true &&
-                    dataElements.values.containsAll(options)
+                    dataElements.values.containsAll(dataValues)
             }.mapNotNull {
                 d2.enrollment("${it.enrollment()}")
             }.map {
@@ -258,35 +249,12 @@ class DataManagerImpl
                     .withTrackedEntityAttributeValues()
                     .one().blockingGet()
 
-                transform(tei, program, it)
+                transformations.transform(tei, program, it)
             },
         )
     }.buffer()
         .conflate()
         .flowOn(Dispatchers.IO)
-
-    override suspend fun trackedEntityInstances(
-        ou: String,
-        program: String,
-    ) = withContext(Dispatchers.IO) {
-        val repository = d2.trackedEntityModule().trackedEntityInstanceQuery()
-
-        return@withContext if (networkUtils.isOnline()) {
-            repository.onlineFirst().allowOnlineCache().eq(true)
-                .byOrgUnits().eq(ou)
-                .byProgram().eq(program)
-                .blockingGet()
-                .flatMap { tei -> listOf(tei) }
-                .map { tei -> transform(tei, program) }
-        } else {
-            repository.offlineOnly().allowOnlineCache().eq(false)
-                .byOrgUnits().eq(ou)
-                .byProgram().eq(program)
-                .blockingGet()
-                .flatMap { tei -> listOf(tei) }
-                .map { tei -> transform(tei, program) }
-        }
-    }
 
     @Throws(IllegalArgumentException::class)
     override suspend fun getAttendanceEvent(
@@ -315,7 +283,7 @@ class DataManagerImpl
                 .withTrackedEntityDataValues()
                 .blockingGet()
                 .mapNotNull {
-                    eventTransform(it, dataElement, reasonDataElement)
+                    transformations.eventTransform(it, dataElement, reasonDataElement)
                 }
                 .map { attendanceEntity ->
                     val status = config.attendanceStatus?.find { status ->
@@ -377,7 +345,7 @@ class DataManagerImpl
                         !cursor.isNull(1) && !cursor.isNull(2)
                     ) {
                         val response = async {
-                            teiEventTransform(
+                            transformations.teiEventTransform(
                                 teiUid = cursor.getString(1),
                                 eventUid = cursor.getString(0),
                                 program = program,
@@ -446,149 +414,5 @@ class DataManagerImpl
                     code = it.code(),
                 )
             }
-    }
-
-    private fun teiEventTransform(
-        teiUid: String,
-        eventUid: String,
-        program: String,
-        attendanceDataElement: String,
-        reasonDataElement: String,
-        config: Attendance,
-    ): Pair<SearchTeiModel, AttendanceEntity> {
-        val tei = d2.trackedEntityModule()
-            .trackedEntityInstances()
-            .byUid().eq(teiUid)
-            .withTrackedEntityAttributeValues()
-            .one().blockingGet()
-
-        val event = d2.eventModule().events()
-            .byUid().eq(eventUid)
-            .withTrackedEntityDataValues()
-            .one().blockingGet()
-        val enrollment = d2.enrollment(event!!.enrollment() ?: "")
-
-        val teiModel = transform(tei, program, enrollment)
-        val transformedEvent = eventTransform(event, attendanceDataElement, reasonDataElement)
-
-        val status = config.attendanceStatus?.find { status ->
-            status.code == transformedEvent?.value
-        }
-
-        val attendanceEntity = transformedEvent?.withBtnSettings(
-            icon = Utils.dynamicIcons("${status?.icon}"),
-            iconName = "${status?.icon}",
-            iconColor = getAttendanceStatusColor("${status?.key}", "${status?.color}"),
-        )
-
-        return Pair(teiModel, attendanceEntity!!)
-    }
-
-    private fun eventTransform(
-        event: Event,
-        dataElement: String,
-        reasonDataElement: String?,
-    ): AttendanceEntity? {
-        val dataValue = event.trackedEntityDataValues()?.find { it.dataElement() == dataElement }
-        val reason = event.trackedEntityDataValues()?.find { it.dataElement() == reasonDataElement }
-
-        return if (dataValue != null) {
-            val tei = d2.enrollment(event.enrollment().toString())?.trackedEntityInstance() ?: ""
-
-            AttendanceEntity(
-                tei = tei,
-                dataElement = dataElement,
-                value = dataValue.value().toString(),
-                reasonDataElement = if (reason == null) {
-                    null
-                } else {
-                    reasonDataElement
-                },
-                reasonOfAbsence = reason?.value(),
-                date = DateHelper.formatDate(
-                    event.eventDate()?.time ?: DateUtils.getInstance().today.time,
-                ).toString(),
-            )
-        } else {
-            null
-        }
-    }
-
-    private fun transform(
-        tei: TrackedEntityInstance?,
-        program: String?,
-        enrollment: Enrollment? = null,
-    ): SearchTeiModel {
-        val searchTei = SearchTeiModel()
-        searchTei.tei = tei
-        currentProgram = program ?: ""
-
-        if (tei?.trackedEntityAttributeValues() != null) {
-            if (program != null) {
-                val programAttributes = d2.programModule().programTrackedEntityAttributes()
-                    .byProgram().eq(program)
-                    .byDisplayInList().isTrue
-                    .orderBySortOrder(RepositoryScope.OrderByDirection.ASC)
-                    .blockingGet()
-
-                for (programAttribute in programAttributes) {
-                    val attribute = d2.trackedEntityModule().trackedEntityAttributes()
-                        .uid(programAttribute.trackedEntityAttribute()!!.uid())
-                        .blockingGet()
-
-                    for (attrValue in tei.trackedEntityAttributeValues()!!) {
-                        if (attrValue.trackedEntityAttribute() == attribute?.uid()) {
-                            addAttribute(searchTei, attrValue, attribute)
-                            break
-                        }
-                    }
-                }
-            } else {
-                val typeAttributes = d2.trackedEntityModule().trackedEntityTypeAttributes()
-                    .byTrackedEntityTypeUid().eq(searchTei.tei.trackedEntityType())
-                    .byDisplayInList().isTrue
-                    .blockingGet()
-                for (typeAttribute in typeAttributes) {
-                    val attribute = d2.trackedEntityModule().trackedEntityAttributes()
-                        .uid(typeAttribute.trackedEntityAttribute()!!.uid())
-                        .blockingGet()
-                    for (attrValue in tei.trackedEntityAttributeValues()!!) {
-                        if (attrValue.trackedEntityAttribute() == attribute?.uid()) {
-                            addAttribute(searchTei, attrValue, attribute)
-                            break
-                        }
-                    }
-                }
-            }
-        }
-
-        if (enrollment != null) {
-            searchTei.addEnrollment(enrollment)
-        }
-
-        searchTei.displayOrgUnit = displayOrgUnit()
-        return searchTei
-    }
-
-    private fun addAttribute(
-        searchTei: SearchTeiModel,
-        attrValue: TrackedEntityAttributeValue,
-        attribute: TrackedEntityAttribute?,
-    ) {
-        val friendlyValue = attrValue.userFriendlyValue(d2)
-
-        val attrValueBuilder = TrackedEntityAttributeValue.builder()
-        attrValueBuilder.value(friendlyValue)
-            .created(attrValue.created())
-            .lastUpdated(attrValue.lastUpdated())
-            .trackedEntityAttribute(attrValue.trackedEntityAttribute())
-            .trackedEntityInstance(searchTei.tei.uid())
-        searchTei.addAttributeValue(attribute?.displayFormName(), attrValueBuilder.build())
-    }
-
-    private fun displayOrgUnit(): Boolean {
-        return d2.organisationUnitModule().organisationUnits()
-            .byProgramUids(listOf(currentProgram))
-            .blockingGet().size > 1
     }
 }
