+++ emis/src/main/java/org/saudigitus/emis/helper/ISEMISSync.kt
@@ -1,0 +1,13 @@
+package org.saudigitus.emis.helper
+
+import org.saudigitus.emis.data.model.SearchTeiModel
+
+interface ISEMISSync {
+
+    suspend fun downloadTEIsByUids(
+        ou: String,
+        program: String,
+        dataElementIds: List<String>,
+        dataValues:List<String>,
+    )
+}
