+++ emis/src/main/java/org/saudigitus/emis/ui/home/HomeUiEvent.kt
@@ -10,4 +10,5 @@ sealed class HomeUiEvent {
     data object OnBack : HomeUiEvent()
     data object HideShowFilter : HomeUiEvent()
     data object Sync : HomeUiEvent()
+    data object OnDownloadStudent: HomeUiEvent()
 }
